Task2/requirements.txt
datasets==3.2.0
evaluate==0.4.3
joblib==1.4.2
numpy==2.2.3
python-dotenv==1.0.1
torch==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.0
transformers==4.48.0
matplotlib==3.10.0